feat(stdlibs/std)!: change TestSetPrevRealm to TestSetRealm (#2164)
This PR removes the `TestSetPrevRealm` and `TestSetPrevAddr`, added in
#890, in favour of a new API `TestSetRealm`, which is more flexible in
its behaviour.

The function works by setting the result of `CurrentRealm()` for the
frame in which the `TestSetRealm` function is called. Consequently,
calling `PrevRealm` in any function called within will yield the value
passed to `TestSetRealm`.

`TestSetRealm` allows to transparently set "user" and "code" realms, by
using two new costructors in the testing stdlibs, `std.NewUserRealm` and
`std.NewCodeRealm`, allowing emulation of both kinds of callers.

The call is not `TestSetPrevRealm` as it does not permanently "set" the
value returned by PrevRealm, as if in a mock; it simply changes the
Realm of the current frame, allowing "deep" calls to PrevRealm() to
correctly return the value of the previous realm.


<details><summary>Contributors' checklist...</summary>

- [ ] Added new tests, or not needed, or not feasible
- [ ] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [ ] Updated the official documentation or not needed
- [ ] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [ ] Added references to related issues and PRs
- [ ] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>