fix(blog): add edit check (#2271)
<!-- please provide a detailed description of the changes made in this
pull request. -->

## Description

This PR provides a hotfix to the edit functionality in the blog package.
This is not the best solution, but due to the urgency of the requests, a
we cannot wait for a full refactor of the code, and there are still
issues with [AVL & GnoVM](https://github.com/gnolang/gno/issues/2266)
that are blocking the other proper workaround on this.

<details><summary>Contributors' checklist...</summary>

- [x] Added new tests, or not needed, or not feasible
- [x] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [x] Updated the official documentation or not needed
- [x] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [x] Added references to related issues and PRs
- [ ] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>