fix(makefile): `gnodev` with `make install` (#2346)
<!-- please provide a detailed description of the changes made in this
pull request. -->

In this PR the Makefile for `contribs/` was changed, so running `make
install` in the root Makefile will not install `gnodev` properly. I've
fixed the path to get back to the original, intended functionality.

<details><summary>Contributors' checklist...</summary>

- [x] Added new tests, or not needed, or not feasible
- [x] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [x] Updated the official documentation or not needed
- [x] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [x] Added references to related issues and PRs
- [ ] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>