chore(pages/events): update events (#2347)
<!-- please provide a detailed description of the changes made in this
pull request. -->

## Description

This PR updates the events page.

![Screenshot 2024-06-12 at 23 57
36](https://github.com/gnolang/gno/assets/33522493/4d2583ac-52ae-47a0-abfd-31de7c73c95d)

Soon, we will have a `r/gnoland/events` realm that will simplify this
process of updating events through the CLI, instead of having to make a
PR every time. There is an open issue on this somewhere but I cannot
seem to find it.

<details><summary>Contributors' checklist...</summary>

- [x] Added new tests, or not needed, or not feasible
- [x] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [x] Updated the official documentation or not needed
- [x] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [x] Added references to related issues and PRs
- [ ] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>