fix(gnovm): source location always precise with line and column
The column number was not stored in the GnoVM location. A `Nonce`
field was used to distinguished 2 block nodes on the same line.
This was applied only to block nodes, not expressions or statements.

As a result, the node location was the same for all expressions and
statements located at the same source line. This change ensures the
uniqueness of location for all nodes, in addition to providing an
accurate position in source code.

We replace `Nonce` by `Column`, computed by the Go tokenizer (the same
as line numbers). Column is always included in all AST nodes, and the
location string has a uniform format.

With this change it is now possible for the debugger to step in
infinite loops on one line like:

`for { i++; println(i) }`

This change is also necessary for a correct accounting for code
coverage, and also may gas accounting.