fix(contribs): revert back to `make <action>.<program>` syntax instead of the opposite (#2353)
Please, can we take a bit of care when modifying the syntaxes of our
makefile commands?

They may not break code, but they break documentation and user flows. We
can add a CI check as well; I see this for another PR as this is a
hotfix. (Currently, running `make install` / `make install.gnodev` in
the top-level directory doesn't work; because @moul inverted these in
#1945). This is still not a substitute for a bit of good ol' attention
in recognising when something is user-facing and thus should be changed
with a bit more care and reason.

Note, I'm NOT saying that we should not change things, as we are before
any 1.0 so we should change anything that makes sense; but it is a good
practice that when changing a flag, command, etc., to do a global search
for the old syntax to see examples where old documentation or untested
code is broken.