feat(keeper)!: use <pkgpath>.<expr> syntax for qeval; <pkgpath>:<path> for qrender (#2382)
This is a switch that was discussed with @leohhhn and @moul. It switches
the current syntax for qeval, which requires a newline as a separator,
to use a dot `.` instead (ie. `gno.land/r/demo/users.MyFunction(123)`).
For qrender, this is switched to a colon `:`, like the gnoweb render:
`gno.land/r/demo/users:u/morgan`.

BREAKING CHANGE: current qeval and qrender calls using the RPC endpoints
will have to be changed. No changes are required for gnoclient users.

<details><summary>Contributors' checklist...</summary>

- [x] Added new tests, or not needed, or not feasible
- [x] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [x] Updated the official documentation or not needed
- [x] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [x] Added references to related issues and PRs
- [x] Provided any useful hints for running manual tests
- [x] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>

---------

Co-authored-by: Manfred Touron <94029+moul@users.noreply.github.com>
Co-authored-by: Leon Hudak <33522493+leohhhn@users.noreply.github.com>