chore(rpcclient): avoid losing jsonrpc information upon non-2xx http codes (fix #2422)
Regardless of HTTP response code, if jsonrpc was received
we want to unmarshal and use it as the response (#2422).

Otherwise we get errors such as:

--= Error =--
Data: unable to call RPC method abci_query, invalid status code received, 500
Msg Traces:
    0  /me/gno/tm2/pkg/crypto/keys/client/query.go:94 - querying
--= /Error =--

When a tool like wireshark shows that the following jsonrpc was actually
in the 500 message:

 JSON compact form: {...}
        "jsonrpc": "2.0"
        "id": ""
        "error": {...}
            "code": -32603
            "message": "Internal error"
            "data": "assignment to entry in nil map"

--------------
Discussion:

An easy way to fix #2422 is to simply append the response body in
case of an HTTP status > 299. But doing so returns an error.
Is this what we want?
---------------------------

The table below indicates what to do based on two bool variables:
 1. is HTTP code 2xx?
 2. can response body be unmarshalled?

/--------+------------------+--------------------------------------\
|        | jsonrpc          | text                                 |
|--------+------------------+--------------------------------------|
| 2xx    | use as response  | Error "unable to unmarshal response" |
| >=300  | use as response⁺ | Error "invalid status code"          |
\------------------------------------------------------------------/
In the case marked with ⁺, we prepend a fmt.Print to show the HTTP
response code to the user.