chore(rpcclient): show body in errors upon non-OK HTTP codes (fix #2422)
When an RPC server responds with a 500 Internal Error,
it sends a jsonrpc with the error that happened.

That information is currently not shown by gnokey however.
Instead an error "invalid status code received, 500" is shown.

This PR adds the body of the response sent by the server to that error.

--= Error =--
Data: unable to call RPC method abci_query, invalid status code received, 500. Response body: "\n
{\"rpcversion\":2,\"id\":\"\",\"error\":{\"code\":-32603,\"message\":\"internal error\"
,\"data\":\"assignment to entry in nil map\"}}"
Msg Traces:
    0  /gno/tm2/pkg/crypto/keys/client/query.go:94 - querying
--= /Error =--