fix(gnovm): source location always precise with line and column (#2362)
The column number was not stored in the GnoVM location. A `Nonce` field
was used to distinguished 2 block nodes on the same line. This was
applied only to block nodes, not expressions or statements.

As a result, the node location was the same for all expressions and
statements located at the same source line. This change ensures the
uniqueness of location for all nodes, in addition to providing an
accurate position in source code.

We replace `Nonce` by `Column`, computed by the Go tokenizer (the same
as line numbers). Column is always included in all AST nodes, and the
location string has a uniform format.

With this change it is now possible for the debugger to step in infinite
loops on one line like:

`for { i++; println(i) }`

This change is also necessary for a correct accounting for code
coverage, and also may be gas accounting.

Note: as `Nonce` is renamed to `Column` in protocol definitions (type
and index unchanged),
it may qualify as a BREAKING CHANGE, but I'm not completely sure of
that.
<!-- please provide a detailed description of the changes made in this
pull request. -->

<details><summary>Contributors' checklist...</summary>

- [*] Added new tests, or not needed, or not feasible
- [*] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [*] Updated the official documentation or not needed
- [*] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [ ] Added references to related issues and PRs
- [*] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>