feat: Bound Method Realm support (#1257)
When a function is called, it's usually enough to get the pkgpath from
where the function is declared, to find out whether we are switching to
a new realm or not.

But when the function is a bound method, and the method and declared
type are defined in a /p/ package, we need to look at the owning
(persisting) pkgpath of the receiver, if any. When one is found, it uses
that.

There are some edge cases to consider, I'm not sure if they are all
fixed, but we will find out.
Here are some that are considered;
1. if the receiver is an unpersisted new object with method, declared in
a p module, it has no realm until it becomes persisted.
2. nil receivers aren't objects, so even if it were stored in a realm,
it would be stored by "reference" from a pointervalue which also is not
an object, so there's no realm to consider even after persistence.
3. if the method from 1 makes any modifications to the /p/ package's
block state, it should result in an error and the tx should fail.
4. if a /p/ declared thing with method gets stored in realm X, it should
be possible for realm Y to call it, and have it modify realm X (e.g. say
realm X attached a closure to the thing).

---------

Signed-off-by: moul <94029+moul@users.noreply.github.com>
Co-authored-by: moul <94029+moul@users.noreply.github.com>