test(gnoclient): Fix a potential bug in the integration test (#2558)
![Screenshot from 2024-07-09
21-48-59](https://github.com/gnolang/gno/assets/167175638/a4429fb7-c277-4296-86fb-f55dfaa924ca)

I'm not sure about the root cause, but it seems like the test node isn't
ready to handle the request call, so it returns an error. Using the
result immediately without asserting that there's no error will cause
this section to panic, leading to a CI that fails randomly

You can see
https://github.com/gnolang/gno/actions/runs/9855164413/job/27209502761?pr=2518

<!-- please provide a detailed description of the changes made in this
pull request. -->

<details><summary>Contributors' checklist...</summary>

- [ ] Added new tests, or not needed, or not feasible
- [ ] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [ ] Updated the official documentation or not needed
- [ ] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [ ] Added references to related issues and PRs
- [ ] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>

---------

Co-authored-by: Morgan <git@howl.moe>