feat: r/profile dapp (#1983)
Following this
[PR](https://github.com/gnolang/gno/pull/181#issuecomment-2061575269)
concerning the creation of a realm `profile` I created this realm which
allows the creation of profile as well as the associated functions to
display the information of a profile with an address or a username. I
have some questions concerning this realm:

- Currently, if a user modifies his username, his old username is not
freed and is therefore no longer available, even if it is no longer in
use. Should I free the old username when the user has changed username?
- To make it possible to search by username and address, I've created a
second avl tree containing both username and address, so that I can find
the profile indexed by its address by searching for it by its username.
This is the most efficient solution I've found. I'd like to get some
feedback on this and know if I should do things differently so that
searching by username is more optimized.
- Do you have any other suggestions for completing the profile fields,
or other interesting features to add?

Thanks in advance for your feedback

<details><summary>Contributors' checklist...</summary>

- [ ] Added new tests, or not needed, or not feasible
- [ ] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [ ] Updated the official documentation or not needed
- [ ] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [ ] Added references to related issues and PRs
- [ ] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>

Closes #181

---------

Co-authored-by: kazai <kazai@777>
Co-authored-by: Manfred Touron <94029+moul@users.noreply.github.com>
Co-authored-by: Leon Hudak <33522493+leohhhn@users.noreply.github.com>