feat(gnovm): enforce formatting of added packages (#2464)
This PR introduces the concept of automatically formatting packages
during AddPkg. The advantage is that it's relatively inexpensive since
we were already parsing the submitted source code AST to apply some
verification.

I believe that having a consistent official formatting for the Go
language is one of the things that makes it better, simpler, and more
readable. I think Gno should strive to be as good as or better than Go
in this regard.

Another justification for this approach is that Gno is not just the
language but a platform (gno.land), like a PAAS, and AddPkg can be
considered the "CD" process, while auto-formatting could be the "CI"
component, which often go hand-in-hand in "CI/CD".

The biggest drawback I can see is that in practice, the code content may
not be the same on the publisher's computer and on-chain.

---------

Signed-off-by: moul <94029+moul@users.noreply.github.com>
Co-authored-by: Morgan <morgan@morganbaz.com>