chore: gnoclient: Require Caller in Msg. Remove syntax sugar Msgs (#2715)
BREAKING CHANGE: The gnoclient API provides "syntax sugar" slices of
messages like MsgCall. This PR removes them and instead uses the vm
message definitions directly (without modification). This requires users
of the gnoclient API to specify all needed fields, including `Caller`
but it also simplifies the Sign logic. The concerns of message creation
and message handling are now separated.

* Remove "syntax sugar" `MsgCall`, `MsgSend`, `MsgRun` and
`MsgAddPackage`.
* Remove the related `validateMsgCall`, `validateMsgSend`,
`validateMsgRun` and `validateMsgAddPackage`.
* In `Call`, etc. change the `msgs` param to use `vm.MsgCall`, etc. Use
the `ValidateBasic` method of these types.
* Do not set the `Caller` from the `Signer`. Expect the `Caller` field
to already be filled in.
* In client_test.go and integration_test.go, specify the `Caller` where
needed.
* In client_test.go error tests, change `expectedError` to a string and
use `assert.ErrorContains`. (This allows mixing errors that are defined
by a type with errors defined as singleton objects.)
* Update contribs/gnodev/pkg/dev/node_state_test.go to use `vm.MsgCall`.

<details><summary>Contributors' checklist...</summary>

- [x] Added new tests, or not needed, or not feasible
- [x] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [x] Updated the official documentation or not needed
- [x] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
</details>

---------

Signed-off-by: Jeff Thompson <jeff@thefirst.org>